put manifest and sw in one file
